add hydrator interface
