+++ baseline.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <files psalm-version="4.19.0@a2ad69ae4f5ab1f7d225a8dc4e2ec2d9415ed599">
-  <file src="src/Serializer/Hydrator.php">
+  <file src="src/Serializer/DefaultHydrator.php">
     <MixedAssignment occurrences="5">
       <code>$data[$propertyMetadata['fieldName']]</code>
       <code>$value</code>
